split test_vrf.rb to test_vrf.rb and test_vrf_af.rb
